backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Cors
PyJWT
prometheus-flask-exporter
psycopg2-binary
